nsc: Fix potential integer overflow
https://hackerone.com/reports/513470

(cherry picked from commit 30e0a2fd82cb53b18fae620e5e4ca3dea03fabe1)

+++ modules/demux/nsc.c
@@ -197,7 +197,7 @@ static char *nscdec( vlc_object_t *p_demux, char* p_encoded )
             msg_Err( p_demux, "load_byte failed" );
             return NULL;
         }
-        length |= tmp << ((i - 1) * 8);
+        length |= (unsigned int)tmp << ((i - 1) * 8);
     }
 
     if( length == 0 )
